# backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CronEnableRequest": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "CronEnableRequest",
        "type": "object"
      },
      "CronJobCreateRequest": {
        "properties": {
          "agent_id": {
            "default": "default",
            "title": "Agent Id",
            "type": "string"
          },
          "channel": {
            "default": "",
            "title": "Channel",
            "type": "string"
          },
          "chat_id": {
            "default": "",
            "title": "Chat Id",
            "type": "string"
          },
          "cron": {
            "description": "Crontab expression, e.g. '*/5 * * * *'",
            "title": "Cron",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "name",
          "cron",
          "prompt"
        ],
        "title": "CronJobCreateRequest",
        "type": "object"
      },
      "CronJobUpdateRequest": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          },
          "cron": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cron"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          }
        },
        "title": "CronJobUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeartbeatUpdateRequest": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "interval_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Minutes"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          }
        },
        "title": "HeartbeatUpdateRequest",
        "type": "object"
      },
...
      "TaskCreateRequest": {
        "properties": {
          "auto_run": {
            "default": true,
            "title": "Auto Run",
            "type": "boolean"
          },
          "gateway": {
            "default": "",
            "title": "Gateway",
            "type": "string"
          },
          "gateway_user": {
            "default": "",
            "title": "Gateway User",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "prompt"
        ],
        "title": "TaskCreateRequest",
        "type": "object"
      },
      "TaskResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "gateway": {
            "default": "",
            "title": "Gateway",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "result": {
            "default": "",
            "title": "Result",
            "type": "string"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "prompt",
          "status",
          "thread_id",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "钻井行业 DeepAgents 智能体系统 - 自主规划与执行",
    "title": "BoetClaw DeepAgents",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/cron": {
      "get": {
        "operationId": "list_cron_jobs_api_v1_tasks_cron_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Cron Jobs",
        "tags": [
          "Scheduling"
        ]
      },
      "post": {
        "operationId": "create_cron_job_api_v1_tasks_cron_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronJobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cron Job",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/tasks/cron/history": {
      "get": {
        "operationId": "cron_history_api_v1_tasks_cron_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cron History",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/tasks/cron/{job_id}": {
      "delete": {
        "operationId": "delete_cron_job_api_v1_tasks_cron__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cron Job",
        "tags": [
          "Scheduling"
        ]
      },
      "put": {
        "operationId": "update_cron_job_api_v1_tasks_cron__job_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronJobUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Cron Job",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/tasks/cron/{job_id}/enable": {
      "post": {
        "operationId": "enable_cron_job_api_v1_tasks_cron__job_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronEnableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Cron Job",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/tasks/cron/{job_id}/trigger": {
      "post": {
        "operationId": "trigger_cron_job_api_v1_tasks_cron__job_id__trigger_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Cron Job",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/tasks/heartbeat": {
      "get": {
        "operationId": "get_heartbeat_api_v1_tasks_heartbeat_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Heartbeat",
        "tags": [
          "Scheduling"
        ]
      },
      "put": {
        "operationId": "update_heartbeat_api_v1_tasks_heartbeat_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Heartbeat",
        "tags": [
          "Scheduling"
        ]
      }
    },
...
  }
}